Removed getters from nodes and refactored to prevent duplication
