+++ spec/Glitch/Grammar/Tree/AddListenerNodeSpec.php
@@ -20,16 +20,6 @@ class AddListenerNodeSpec extends ObjectBehavior
         $this->shouldHaveType('Glitch\Grammar\Tree\StatementNode');
     }
 
-    function it_has_a_left_hand_side(ReferenceNode $left)
-    {
-        $this->getLeft()->shouldBe($left);
-    }
-
-    function it_has_a_right_hand_side(ReferenceNode $right)
-    {
-        $this->getRight()->shouldBe($right);
-    }
-
     function it_adds_a_listener_when_invoked(ActivationObject $scope, ReferenceNode $left, ReferenceNode $right, EventValue $event, EventValue $listener)
     {
         $left->reduce($scope)->willReturn($event);
