+++ spec/Glitch/Grammar/Tree/RemoveListenerNodeSpec.php
@@ -14,13 +14,8 @@ class RemoveListenerNodeSpec extends ObjectBehavior
         $this->beConstructedWith($left, $right);
     }
 
-    function it_has_a_left_hand_side(ReferenceNode $left)
+    function it_is_a_statement()
     {
-        $this->getLeft()->shouldBe($left);
-    }
-
-    function it_has_a_right_hand_side(StringNode $right)
-    {
-        $this->getRight()->shouldBe($right);
+        $this->shouldHaveType('Glitch\Grammar\Tree\StatementNode');
     }
 }
