+++ src/Glitch/Grammar/Tree/AssignmentNode.php
@@ -2,24 +2,12 @@
 
 namespace Glitch\Grammar\Tree;
 
-class AssignmentNode
+use Glitch\Interpreter\ActivationObject;
+
+class AssignmentNode extends StatementNode
 {
-    private $left;
-    private $right;
-
-    public function __construct($left, $right)
+    public function invoke(ActivationObject $scope)
     {
-        $this->left = $left;
-        $this->right = $right;
-    }
-
-    public function getLeft()
-    {
-        return $this->left;
-    }
-
-    public function getRight()
-    {
-        return $this->right;
+        // TODO
     }
 }
