+++ src/Glitch/Grammar/Tree/NullaryExpressionNode.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Glitch\Grammar\Tree;
+
+abstract class NullaryExpressionNode implements ExpressionNode
+{
+    protected $value;
+
+    public function __construct($value)
+    {
+        $this->value = $value;
+    }
+}
+
